pyproject.toml
[project]
name = "abidpa02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "chainlit>=2.6.9",
    "fastapi>=0.116.1",
    "openai-agents>=0.2.8",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
